+++ descriptor/way_obj_desc.h
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 1997 - 2002 by Volker Meyer & Hansjörg Malthaner
+ *  Copyright (c) 1997 - 2002 by Volker Meyer & Hansjï¿½rg Malthaner
  *
  * This file is part of the Simutrans project under the artistic licence.
  *
@@ -61,7 +61,7 @@ public:
 
 	image_id get_front_slope_image_id(slope_t::type slope) const
 	{
-		int nr;
+		uint16 nr;
 		switch(slope) {
 			case 4:
 				nr = 0;
@@ -101,7 +101,7 @@ public:
 
 	image_id get_back_slope_image_id(slope_t::type slope) const
 	{
-		int nr;
+		uint16 nr;
 		switch(slope) {
 			case 4:
 				nr = 0;
